feat(page): move page.pdf to PDF streams
This lets transferring massive PDF files over the protocol.

Fix #4563
