+++ lib/Page.js
@@ -924,6 +924,7 @@ class Page extends EventEmitter {
     const marginRight = convertPrintParameterToInches(margin.right) || 0;
 
     const result = await this._client.send('Page.printToPDF', {
+      transferMode: 'ReturnAsStream',
       landscape,
       displayHeaderFooter,
       headerTemplate,
@@ -939,10 +940,7 @@ class Page extends EventEmitter {
       pageRanges,
       preferCSSPageSize
     });
-    const buffer = Buffer.from(result.data, 'base64');
-    if (path !== null)
-      await writeFileAsync(path, buffer);
-    return buffer;
+    return await helper.readProtocolStream(this._client, result.stream, path);
   }
 
   /**
