+++ lib/Tracing.js
@@ -14,11 +14,6 @@
  * limitations under the License.
  */
 const {helper, assert} = require('./helper');
-const fs = require('fs');
-
-const openAsync = helper.promisify(fs.open);
-const writeAsync = helper.promisify(fs.write);
-const closeAsync = helper.promisify(fs.close);
 
 class Tracing {
   /**
@@ -66,39 +61,11 @@ class Tracing {
     let fulfill;
     const contentPromise = new Promise(x => fulfill = x);
     this._client.once('Tracing.tracingComplete', event => {
-      this._readStream(event.stream, this._path).then(fulfill);
+      helper.readProtocolStream(this._client, event.stream, this._path).then(fulfill);
     });
     await this._client.send('Tracing.end');
     this._recording = false;
     return contentPromise;
-  }
-
-  /**
-   * @param {string} handle
-   * @param {?string} path
-   */
-  async _readStream(handle, path) {
-    let eof = false;
-    let file;
-    if (path)
-      file = await openAsync(path, 'w');
-    const bufs = [];
-    while (!eof) {
-      const response = await this._client.send('IO.read', {handle});
-      eof = response.eof;
-      bufs.push(Buffer.from(response.data));
-      if (path)
-        await writeAsync(file, response.data);
-    }
-    if (path)
-      await closeAsync(file);
-    await this._client.send('IO.close', {handle});
-    let resultBuffer = null;
-    try {
-      resultBuffer = Buffer.concat(bufs);
-    } finally {
-      return resultBuffer;
-    }
   }
 }
